backend/gemini_alpaca_agent/requirements.txt
google-genai>=1.68.0
alpaca-py>=0.42.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
